Move tabor segment analysis to _program subpackage
