+++ tests/hardware/tabor_dummy_based_tests.py
@@ -461,97 +461,6 @@ class TaborChannelPairTests(TaborDummyBasedTest):
                                                            voltage_transformations=test_transform)
                 self.assertEqual([], offset_mock.call_args_list)
 
-    def test_find_place_for_segments_in_memory(self):
-        def hash_based_on_dir(ch):
-            hash_list = []
-            for d in dir(ch):
-                o = getattr(ch, d)
-                if isinstance(o, np.ndarray):
-                    hash_list.append(hash(o.tobytes()))
-                else:
-                    try:
-                        hash_list.append(hash(o))
-                    except TypeError:
-                        pass
-            return hash(tuple(hash_list))
-
-        channel_pair = TaborChannelPair(self.instrument, identifier='asd', channels=(1, 2))
-
-        # empty
-        segments = np.asarray([-5, -6, -7, -8, -9])
-        segment_lengths = 192 + np.asarray([32, 16, 64, 32, 16])
-
-        hash_before = hash_based_on_dir(channel_pair)
-
-        w2s, ta, ti = channel_pair._find_place_for_segments_in_memory(segments, segment_lengths)
-        self.assertEqual(w2s.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(ta.tolist(), [True, True, True, True, True])
-        self.assertEqual(ti.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(hash_before, hash_based_on_dir(channel_pair))
-
-        # all new segments
-        channel_pair._segment_capacity = 192 + np.asarray([0, 16, 32, 16, 0], dtype=np.uint32)
-        channel_pair._segment_hashes = np.asarray([1, 2, 3, 4, 5], dtype=np.int64)
-        channel_pair._segment_references = np.asarray([1, 1, 1, 2, 1], dtype=np.int32)
-        hash_before = hash_based_on_dir(channel_pair)
-
-        w2s, ta, ti = channel_pair._find_place_for_segments_in_memory(segments, segment_lengths)
-        self.assertEqual(w2s.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(ta.tolist(), [True, True, True, True, True])
-        self.assertEqual(ti.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(hash_before, hash_based_on_dir(channel_pair))
-
-        # some known segments
-        channel_pair._segment_capacity = 192 + np.asarray([0, 16, 32, 64, 0, 16], dtype=np.uint32)
-        channel_pair._segment_hashes = np.asarray([1, 2, 3, -7, 5, -9], dtype=np.int64)
-        channel_pair._segment_references = np.asarray([1, 1, 1, 2, 1, 3], dtype=np.int32)
-        hash_before = hash_based_on_dir(channel_pair)
-
-        w2s, ta, ti = channel_pair._find_place_for_segments_in_memory(segments, segment_lengths)
-        self.assertEqual(w2s.tolist(), [-1, -1, 3, -1, 5])
-        self.assertEqual(ta.tolist(), [True, True, False, True, False])
-        self.assertEqual(ti.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(hash_before, hash_based_on_dir(channel_pair))
-
-        # insert some segments with same length
-        channel_pair._segment_capacity = 192 + np.asarray([0, 16, 32, 64, 0, 16], dtype=np.uint32)
-        channel_pair._segment_hashes = np.asarray([1, 2, 3, 4, 5, 6], dtype=np.int64)
-        channel_pair._segment_references = np.asarray([1, 0, 1, 0, 1, 3], dtype=np.int32)
-        hash_before = hash_based_on_dir(channel_pair)
-
-        w2s, ta, ti = channel_pair._find_place_for_segments_in_memory(segments, segment_lengths)
-        self.assertEqual(w2s.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(ta.tolist(), [True, False, False, True, True])
-        self.assertEqual(ti.tolist(), [-1, 1, 3, -1, -1])
-        self.assertEqual(hash_before, hash_based_on_dir(channel_pair))
-
-        # insert some segments with smaller length
-        channel_pair._segment_capacity = 192 + np.asarray([0, 80, 32, 64, 96, 16], dtype=np.uint32)
-        channel_pair._segment_hashes = np.asarray([1, 2, 3, 4, 5, 6], dtype=np.int64)
-        channel_pair._segment_references = np.asarray([1, 0, 1, 1, 0, 3], dtype=np.int32)
-        hash_before = hash_based_on_dir(channel_pair)
-
-        w2s, ta, ti = channel_pair._find_place_for_segments_in_memory(segments, segment_lengths)
-        self.assertEqual(w2s.tolist(), [-1, -1, -1, -1, -1])
-        self.assertEqual(ta.tolist(), [True, True, False, False, True])
-        self.assertEqual(ti.tolist(), [-1, -1, 4, 1, -1])
-        self.assertEqual(hash_before, hash_based_on_dir(channel_pair))
-
-        # mix everything
-        segments = np.asarray([-5, -6, -7, -8, -9, -10, -11])
-        segment_lengths = 192 + np.asarray([32, 16, 64, 32, 16, 0, 0])
-
-        channel_pair._segment_capacity = 192 + np.asarray([0, 80, 32, 64, 32, 16], dtype=np.uint32)
-        channel_pair._segment_hashes = np.asarray([1, 2, 3, 4, -8, 6], dtype=np.int64)
-        channel_pair._segment_references = np.asarray([1, 0, 1, 0, 1, 0], dtype=np.int32)
-        hash_before = hash_based_on_dir(channel_pair)
-
-        w2s, ta, ti = channel_pair._find_place_for_segments_in_memory(segments, segment_lengths)
-        self.assertEqual(w2s.tolist(), [-1,    -1,   -1,    4,     -1,     -1, -1])
-        self.assertEqual(ta.tolist(),  [False, True, False, False, True, True, True])
-        self.assertEqual(ti.tolist(),  [1,     -1,   3,     -1,    -1,   -1,   -1])
-        self.assertEqual(hash_before, hash_based_on_dir(channel_pair))
-
     def test_upload_segment(self):
         channel_pair = TaborChannelPair(self.instrument, identifier='asd', channels=(1, 2))
